requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0
PyJWT==2.8.0
requests==2.32.3
